# public/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Mobile Handset Catalogue API",
        "version": "1.0.0",
        "description": "API documentation for the mobile handset catalogue. Endpoints are grouped by version."
    },
    "tags": [
        {
            "name": "v1",
            "description": "Version 1 endpoints (original response format)"
        },
        {
            "name": "v2",
            "description": "Version 2 endpoints (breaking changes, improved format)"
        }
    ],
    "paths": {
        "/api/v1/handsets": {
            "get": {
                "tags": [
                    "v1"
                ],
                "summary": "List all handsets (v1)",
                "description": "Original response format",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Brand name"
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "Minimum price"
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "Maximum price"
                    },
                    {
                        "name": "release_year",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Release year"
                    },
                    {
                        "name": "features",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "style": "form",
                        "explode": true,
                        "description": "Features (array)"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search text"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        },
                        "description": "Results per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of handsets (v1 format)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "brand": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    },
                                                    "release_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "features": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "specifications": {
                                                        "type": "object"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "filters_applied": {
                                                    "type": "object"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/handsets": {
            "get": {
                "tags": [
                    "v2"
                ],
                "summary": "List all handsets (v2)",
                "description": "Response format with breaking changes (nested price, nested brand, new pagination)",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Brand name"
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "Minimum price"
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "description": "Maximum price"
                    },
                    {
                        "name": "release_year",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Release year"
                    },
                    {
                        "name": "features",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "style": "form",
                        "explode": true,
                        "description": "Features (array)"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search text"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        },
                        "description": "Results per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of handsets (v2 format)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "brand": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "country": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "price": {
                                                        "type": "object",
                                                        "properties": {
                                                            "amount": {
                                                                "type": "number"
                                                            },
                                                            "currency": {
                                                                "type": "string"
                                                            },
                                                            "discount_percentage": {
                                                                "type": "number"
                                                            },
                                                            "final_price": {
                                                                "type": "number"
                                                            }
                                                        }
                                                    },
                                                    "release_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "features": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "specifications": {
                                                        "type": "object"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "pagination": {
                                            "type": "object",
                                            "properties": {
                                                "total_items": {
                                                    "type": "integer"
                                                },
                                                "items_per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                },
                                                "links": {
                                                    "type": "object",
                                                    "properties": {
                                                        "first": {
                                                            "type": "string"
                                                        },
                                                        "last": {
                                                            "type": "string"
                                                        },
                                                        "next": {
                                                            "type": [
                                                                "string",
                                                                "null"
                                                            ]
                                                        },
                                                        "prev": {
                                                            "type": [
                                                                "string",
                                                                "null"
                                                            ]
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}